move lhotse_dataset code to lhotse_s2s_dataset and recover stt dataset
